Change parseInt() to parseFloat()

+++ resources/views/orders/index.blade.php
@@ -2216,7 +2216,7 @@
                                 return 0;
                             }
 
-                            const Z = (sumOfSelling + parseInt(row.additional_cash_on_delivery_cost) - (sumOfPurchase * 1.23)).toFixed(2);
+                            const Z = (sumOfSelling + parseFloat(row.additional_cash_on_delivery_cost) - (sumOfPurchase * 1.23)).toFixed(2);
 
                             const BZO = (parseInt(Z) + parseInt(RKTBO) + parseInt(PSIK) - parseInt(PSW) + parseInt(WAC) + parseFloat(ZP)).toFixed(2);
